requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5
defusedxml==0.5.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11
django-allauth==0.35.0
django-ckeditor==5.4.0
django-js-asset==1.0.0
django-rest-swagger==2.1.2
django-toolbelt==0.0.1
django-wysiwyg==0.8.0
djangorestframework==3.6.3
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
oauthlib==2.0.6
olefile==0.45.1
openapi-codec==1.3.2
Pillow==4.1.1
psycopg2==2.7.1
PyJWT==1.5.3
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2017.3
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.13.2
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
static3==0.7.0
